Mistake in "first_coin_flip" formula for "Varying Amounts of Noise"
There is mistake in this line:    first_coin_flip = (torch.rand(len(db)) *<* noise).float()
it should be:  first_coin_flip = (torch.rand(len(db)) *>* noise).float()
E.g. if noise is 0.2 it means less noise, and when it is 0.8 it means more noise. right?
So, when it's 0.8 I have to get a true result wit probability 1-0.8 = 0.2.